Ensure we do not fail if there is no java processes running

+++ kill_cluster.sh
@@ -1,5 +1,8 @@
 echo "Killing all java processes..."
 ps ax | grep java | grep -v grep | awk '{print $1}'
-ps ax | grep java | grep -v grep | awk '{print $1}' | sudo xargs kill -9
-sleep 3
-ps ax | grep java | grep -v grep | awk '{print $1}' | wc -l
+numberOfProcesses=$(ps ax | grep java | grep -v grep | awk '{print $1}' | wc -l)
+if [ $numberOfProcesses -gt 0 ] ; then
+	ps ax | grep java | grep -v grep | awk '{print $1}' | sudo xargs kill -9
+	sleep 3
+	ps ax | grep java | grep -v grep | awk '{print $1}' | wc -l
+fi
